+++ test/interfaces/PolyNetworkLibraries/Utils.sol
@@ -1,8 +1,6 @@
 pragma solidity ^0.8.0;
 
-
 library Utils {
-
     /* @notice      Convert the bytes array to bytes32 type, the bytes array length must be 32
     *  @param _bs   Source bytes array
     *  @return      bytes32
@@ -10,7 +8,8 @@ library Utils {
     function bytesToBytes32(bytes memory _bs) internal pure returns (bytes32 value) {
         require(_bs.length == 32, "bytes length is not 32.");
         assembly {
-            // load 32 bytes from memory starting from position _bs + 0x20 since the first 0x20 bytes stores _bs length
+            // load 32 bytes from memory starting from position _bs + 0x20 since the first 0x20 bytes stores
+            // _bs length
             value := mload(add(_bs, 0x20))
         }
     }
@@ -25,7 +24,10 @@ library Utils {
             // load 32 bytes from memory starting from position _bs + 32
             value := mload(add(_bs, 0x20))
         }
-        require(value <= 0x7fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff, "Value exceeds the range");
+        require(
+            value <= 0x7fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff,
+            "Value exceeds the range"
+        );
     }
 
     /* @notice      Convert uint256 to bytes
@@ -33,7 +35,10 @@ library Utils {
     *  @return      bytes
     */
     function uint256ToBytes(uint256 _value) internal pure returns (bytes memory bs) {
-        require(_value <= 0x7fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff, "Value exceeds the range");
+        require(
+            _value <= 0x7fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff,
+            "Value exceeds the range"
+        );
         assembly {
             // Get a location of some free memory and store it in result as
             // Solidity does for memory variables.
@@ -51,22 +56,21 @@ library Utils {
     *  @param _bs   Source bytes: bytes length must be 20
     *  @return      Converted address from source bytes
     */
-    function bytesToAddress(bytes memory _bs) internal pure returns (address addr)
-    {
+    function bytesToAddress(bytes memory _bs) internal pure returns (address addr) {
         require(_bs.length == 20, "bytes length does not match address");
         assembly {
             // for _bs, first word store _bs.length, second word store _bs.value
-            // load 32 bytes from mem[_bs+20], convert it into Uint160, meaning we take last 20 bytes as addr (address).
+            // load 32 bytes from mem[_bs+20], convert it into Uint160, meaning we take last 20 bytes as addr
+            // (address).
             addr := mload(add(_bs, 0x14))
         }
-
     }
-    
+
     /* @notice      Convert address to bytes
     *  @param _addr Address need to be converted
     *  @return      Converted bytes from address
     */
-    function addressToBytes(address _addr) internal pure returns (bytes memory bs){
+    function addressToBytes(address _addr) internal pure returns (bytes memory bs) {
         assembly {
             // Get a location of some free memory and store it in result as
             // Solidity does for memory variables.
@@ -77,14 +81,14 @@ library Utils {
             mstore(add(bs, 0x20), shl(96, _addr))
             // Update the free-memory pointer by padding our last write location to 32 bytes
             mstore(0x40, add(bs, 0x40))
-       }
+        }
     }
 
     /* @notice          Do hash leaf as the multi-chain does
     *  @param _data     Data in bytes format
     *  @return          Hashed value in bytes32 format
     */
-    function hashLeaf(bytes memory _data) internal pure returns (bytes32 result)  {
+    function hashLeaf(bytes memory _data) internal pure returns (bytes32 result) {
         result = sha256(abi.encodePacked(bytes1(0x0), _data));
     }
 
@@ -93,7 +97,7 @@ library Utils {
     *  @param _r        Right node
     *  @return          Hashed value in bytes32 format
     */
-    function hashChildren(bytes32 _l, bytes32  _r) internal pure returns (bytes32 result)  {
+    function hashChildren(bytes32 _l, bytes32 _r) internal pure returns (bytes32 result) {
         result = sha256(abi.encodePacked(bytes1(0x01), _l, _r));
     }
 
@@ -124,7 +128,8 @@ library Utils {
             case 1 {
                 // fslot can contain both the length and contents of the array
                 // if slength < 32 bytes so let's prepare for that
-                // v. http://solidity.readthedocs.io/en/latest/miscellaneous.html#layout-of-state-variables-in-storage
+                // v.
+                // http://solidity.readthedocs.io/en/latest/miscellaneous.html#layout-of-state-variables-in-storage
                 // slength != 0
                 if iszero(iszero(slength)) {
                     switch lt(slength, 32)
@@ -182,11 +187,7 @@ library Utils {
     *  @param _length       The index of _bytes for the end of sliced bytes
     *  @return              The sliced bytes
     */
-    function slice(
-        bytes memory _bytes,
-        uint _start,
-        uint _length
-    )
+    function slice(bytes memory _bytes, uint256 _start, uint256 _length)
         internal
         pure
         returns (bytes memory)
@@ -247,16 +248,21 @@ library Utils {
 
         return tempBytes;
     }
+
     /* @notice              Check if the elements number of _signers within _keepers array is no less than _m
     *  @param _keepers      The array consists of serveral address
     *  @param _signers      Some specific addresses to be looked into
     *  @param _m            The number requirement paramter
     *  @return              True means containment, false meansdo do not contain.
     */
-    function containMAddresses(address[] memory _keepers, address[] memory _signers, uint _m) internal pure returns (bool){
-        uint m = 0;
-        for(uint i = 0; i < _signers.length; i++){
-            for (uint j = 0; j < _keepers.length; j++) {
+    function containMAddresses(address[] memory _keepers, address[] memory _signers, uint256 _m)
+        internal
+        pure
+        returns (bool)
+    {
+        uint256 m = 0;
+        for (uint256 i = 0; i < _signers.length; i++) {
+            for (uint256 j = 0; j < _keepers.length; j++) {
                 if (_signers[i] == _keepers[j]) {
                     m++;
                     delete _keepers[j];
@@ -271,19 +277,20 @@ library Utils {
     *  @return
     */
     function compressMCPubKey(bytes memory key) internal pure returns (bytes memory newkey) {
-         require(key.length >= 67, "key lenggh is too short");
-         newkey = slice(key, 0, 35);
-         if (uint8(key[66]) % 2 == 0){
-             newkey[2] = bytes1(0x02);
-         } else {
-             newkey[2] = bytes1(0x03);
-         }
-         return newkey;
+        require(key.length >= 67, "key lenggh is too short");
+        newkey = slice(key, 0, 35);
+        if (uint8(key[66]) % 2 == 0) {
+            newkey[2] = bytes1(0x02);
+        } else {
+            newkey[2] = bytes1(0x03);
+        }
+        return newkey;
     }
-    
+
     /**
      * @dev Returns true if `account` is a contract.
-     *      Refer from https://github.com/OpenZeppelin/openzeppelin-contracts/blob/master/contracts/utils/Address.sol#L18
+     *      Refer from
+     * https://github.com/OpenZeppelin/openzeppelin-contracts/blob/master/contracts/utils/Address.sol#L18
      *
      * This test is non-exhaustive, and there may be false-negatives: during the
      * execution of a contract's constructor, its address will be reported as
