+++ test/utils/IPancakeRouter02.sol
@@ -3,10 +3,10 @@ pragma solidity ^0.8.17;
 
 interface IPancakeRouter02 {
     function swapExactTokensForTokensSupportingFeeOnTransferTokens(
-        uint amountIn,
-        uint amountOutMin,
+        uint256 amountIn,
+        uint256 amountOutMin,
         address[] calldata path,
         address to,
-        uint deadline
+        uint256 deadline
     ) external;
 }
